+++ tools/xslt/Get-XSLTExecutable.ps1
@@ -32,7 +32,7 @@ Param(
 
 )
 
-Function Write-CompilerWarningAndErrors
+Function Write-CompilerWarningAndError
 {
 	[CmdletBinding()]
 	Param(
@@ -173,7 +173,7 @@ try
 				$saxPackage = $saxCompiler.CompilePackage(
 					( [System.IO.FileStream]::new( $XSLTPackagePath, 'Open' ) )
 				);
-				Write-CompilerWarningAndErrors -ErrorList ( $saxCompiler.ErrorList ) `
+				Write-CompilerWarningAndError -ErrorList ( $saxCompiler.ErrorList ) `
 					-ModuleUri $XSLTPackagePath `
 					-Verbose:( $PSCmdlet.MyInvocation.BoundParameters['Verbose'] -eq $true );
 				if ( $PSCmdlet.ShouldProcess( $XSLTPackagePath, 'Import XSLT package' ) )
@@ -183,7 +183,7 @@ try
 			}
 			catch
 			{
-				Write-CompilerWarningAndErrors -ErrorList ( $saxCompiler.ErrorList ) `
+				Write-CompilerWarningAndError -ErrorList ( $saxCompiler.ErrorList ) `
 					-ModuleUri $XSLTPackagePath `
 					-Verbose:( $PSCmdlet.MyInvocation.BoundParameters['Verbose'] -eq $true );
 				$PScmdlet.ThrowTerminatingError();
@@ -204,13 +204,13 @@ try
 		try
 		{
 			$saxExecutable = $saxCompiler.Compile( $LiteralPath );
-			Write-CompilerWarningAndErrors -ErrorList ( $saxCompiler.ErrorList ) `
+			Write-CompilerWarningAndError -ErrorList ( $saxCompiler.ErrorList ) `
 				-ModuleUri $LiteralPath `
 				-Verbose:( $PSCmdlet.MyInvocation.BoundParameters['Verbose'] -eq $true );
 		}
 		catch
 		{
-			Write-CompilerWarningAndErrors -ErrorList ( $saxCompiler.ErrorList ) `
+			Write-CompilerWarningAndError -ErrorList ( $saxCompiler.ErrorList ) `
 				-ModuleUri $LiteralPath `
 				-Verbose:( $PSCmdlet.MyInvocation.BoundParameters['Verbose'] -eq $true );
 			$PScmdlet.ThrowTerminatingError();