+++ CHANGELOG.md
@@ -16,6 +16,9 @@
   (GitHub action 'Release to GihHub')
 - устранено замечание по качеству кода Update-FileLastWriteTime.ps1:
   "Command accepts pipeline input but has not defined a process block".
+- устранено замечание по качеству кода:
+  "The cmdlet 'Write-CompilerWarningAndErrors' uses a plural noun.
+  A singular noun should be used instead".
 
 ## [2.9.0]